Removed nlohmann_json from public headers and fixed install interface
- Moved SimpleTree::node_t definition to SimpleTree.cpp so that
  SimpleTree.h no longer depends on nlohmann_json (avoids leaking
  the dependency to consumers of the installed library)
- Added Rule of Five to SimpleTree: copy disabled, move noexcept;
  destructor and move assignment defined in .cpp where node_t is complete
- Added INSTALL_INTERFACE include dirs to strategy/strategyStatic targets
  so installed consumers find libstrategy headers via cmake
- Fixed FIND_PACKAGE_ARGS version (1.1.2 -> 1.0.9) in libstrategy.cmake
  so the example uses the locally installed library instead of fetching
- Removed nlohmann_json from example/CMakeLists.txt (no longer needed)
