+++ example/CMakeLists.txt
@@ -1,7 +1,6 @@
 cmake_minimum_required(VERSION 3.28)
 project(example VERSION 1.0.9)
 
-include(nlohmann_json)
 include(ptrie)
 include(libstrategy)
