+++ src/CMakeLists.txt
@@ -1,16 +1,22 @@
+include(GNUInstallDirs)
+
 add_library(SimpleTree OBJECT SimpleTree.cpp)
 target_link_libraries(SimpleTree PUBLIC ptrie::ptrie PRIVATE nlohmann_json::nlohmann_json)
 add_library(libstrategy::SimpleTree ALIAS SimpleTree)
 
 add_library(strategy SHARED libz2s.cpp ZonotopStrategy.cpp $<TARGET_OBJECTS:SimpleTree>)
 target_link_libraries(strategy PUBLIC ptrie::ptrie)
-target_include_directories(strategy PRIVATE $<TARGET_PROPERTY:nlohmann_json::nlohmann_json,INTERFACE_INCLUDE_DIRECTORIES>)
+target_include_directories(strategy
+    PUBLIC $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
+    PRIVATE $<TARGET_PROPERTY:nlohmann_json::nlohmann_json,INTERFACE_INCLUDE_DIRECTORIES>)
 set_target_properties(strategy PROPERTIES VISIBILITY_INLINE_HIDDEN ON)
 add_library(libstrategy::strategy ALIAS strategy)
 
 add_library(strategyStatic STATIC libz2s.cpp ZonotopStrategy.cpp $<TARGET_OBJECTS:SimpleTree>)
 target_link_libraries(strategyStatic PUBLIC ptrie::ptrie)
-target_include_directories(strategyStatic PRIVATE $<TARGET_PROPERTY:nlohmann_json::nlohmann_json,INTERFACE_INCLUDE_DIRECTORIES>)
+target_include_directories(strategyStatic
+    PUBLIC $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
+    PRIVATE $<TARGET_PROPERTY:nlohmann_json::nlohmann_json,INTERFACE_INCLUDE_DIRECTORIES>)
 set_target_properties(strategyStatic PROPERTIES OUTPUT_NAME strategy)
 add_library(libstrategy::strategyStatic ALIAS strategyStatic)
 
@@ -19,7 +25,6 @@ if(NOT LIBSTRATEGY_OnlyLibrary)
     target_link_libraries(z2s PUBLIC SimpleTree Boost::program_options)
 endif()
 
-include(GNUInstallDirs)
 include(CMakePackageConfigHelpers)
 
 set(CONFIG_INSTALL_DIR ${CMAKE_INSTALL_LIBDIR}/cmake/libstrategy)
