+++ src/SimpleTree.cpp
@@ -29,10 +29,81 @@
 #include <nlohmann/json.hpp>
 
 #include <iostream>
+#include <limits>
+#include <memory>
+#include <set>
+#include <unordered_map>
 #include <unordered_set>
+#include <utility>
 #include <vector>
 
 using json = nlohmann::json;
+
+struct SimpleTree::node_t : std::enable_shared_from_this<SimpleTree::node_t>
+{
+    uint32_t _var = std::numeric_limits<uint32_t>::max();
+    double _limit = -std::numeric_limits<double>::infinity();
+    double _cost = std::numeric_limits<double>::infinity();
+    std::pair<double, double> _cost_bounds;
+    node_ptr _low;
+    node_ptr _high;
+    node_t* _parent{nullptr};
+    void insert(std::vector<double>& key, json& tree, size_t action, SimpleTree& parent, size_t prefix, bool minimize,
+                double accuracy, std::vector<double>& exactness);
+    std::ostream& print(std::ostream& out, size_t tabs = 0) const;
+    bool is_leaf() const { return _low == nullptr && _high == nullptr; }
+    node_ptr simplify(bool make_dd, nodemap_t& nodemap, SimpleTree& parent);
+    void subsumption_reduction(bool minimization, SimpleTree& parent);
+    void action_nodes(std::vector<node_ptr>& nodes, uint32_t low, uint32_t high, uint32_t varid);
+    std::pair<double, double> compute_min_max() const;
+    bool check_tiles(node_t* start, std::vector<node_ptr>&, std::vector<std::pair<double, double>>& bounds, double val,
+                     double minval, double maxval, bool minimization, size_t offset);
+    bool subsumes(const std::vector<std::pair<double, double>>& bounds, std::vector<std::pair<double, double>>& obounds,
+                  double val, bool minimization, size_t offset, double& best, std::pair<double, double>& closest) const;
+    void get_ranks(std::set<std::pair<double, node_t*>>& values, node_t* start);
+    void set_ranks(std::unordered_map<double, double>& values);
+    std::ostream& print_c(std::ostream& os, size_t disc, std::unordered_set<const node_t*>& printed,
+                          size_t tabs = 0) const;
+    std::ostream& print_c_nested(std::ostream& os, size_t disc, size_t tabs, std::vector<const node_t*>& toprint,
+                                 const node_ptr& node) const;
+    size_t depth() const;
+    double value(const double* disc, const double* cont, uint32_t action, size_t ndisc) const;
+    void consistent(size_t) const;
+    bool cost_intersect(const node_t& other) const;
+    double midcost(const node_t& other, double minval, double maxval) const;
+    bool operator==(const node_t& other) const
+    {
+        if (_limit != other._limit)
+            return false;
+        if (_var != other._var)
+            return false;
+        if (_low != other._low)
+            return false;
+        if (_high != other._high)
+            return false;
+        return true;
+    }
+    bool operator<(const node_t& other) const
+    {
+        if (_limit != other._limit)
+            return _limit < other._limit;
+        if (_var != other._var)
+            return _var < other._var;
+        if (_low != other._low)
+            return _low < other._low;
+        return _high < other._high;
+    }
+    node_ptr& operator[](bool b) { return b ? _high : _low; }
+    operator signature_t() const
+    {
+        if (is_leaf())
+            return signature_t{_var, _limit, _low.get(), _high.get()};
+        return signature_t{0, _cost, nullptr, nullptr};
+    }
+};
+
+SimpleTree& SimpleTree::operator=(SimpleTree&&) noexcept = default;
+SimpleTree::~SimpleTree() = default;
 
 const std::vector<std::string>& SimpleTree::actions() const { return _actions; }
